requirements.txt
mysql-connector-python
bottle
